refactor: factorize some repeated code in various files
- Common code extracted to resilient.utils
- Some SILE hacks extracted to resilient.hacks
